requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.7
distro==1.9.0
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
joblib==1.4.2
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==1.26.4
openai==1.30.5
pandas==2.2.2
pdf2image==1.17.0
pdfminer.six==20231228
pillow==10.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rich==13.7.1
scikit-learn==1.5.0
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
threadpoolctl==3.5.0
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
